requirements.txt
streamlit==1.31.1
openai==1.52.2
streamlit-extras==0.4.0
python-dotenv 
plotly
scipy
scikit-learn
faiss-cpu
pyodbc==4.0.35
SQLAlchemy==2.0.20
kaleido
statsmodels
prophet==1.1.5
statsmodels==0.14.1
tabulate==0.9.0
azure-search-documents==11.4.0
azure-identity
langchain==0.3.4
langchain-openai==0.2.3
langchain-community
langgraph
langsmith
pydantic==2.7.4
geopy
openmeteo-requests
requests-cache
retry-requests
yfinance
httpx==0.27.2
ipython
streamlit-mermaid
seaborn
neo4j
google-api-python-client
azure-ai-inference
nasapy
azure-storage-file-datalake
pypdf
azure-cognitiveservices-speech
mcp
langchain-mcp-adapters

